// libs/hdf-converters/types/splunk-config-types.d.ts
export type SplunkConfig = {
    scheme: string;
    host: string;
    port?: number;
    username?: string;
    password?: string;
    index: string;
    owner?: string;
    app?: string;
    sessionKey?: string;
    autologin?: boolean;
    version?: string;
};
//# sourceMappingURL=splunk-config-types.d.ts.map